Add action syncing for server

+++ server/test/utils.ts
@@ -1,3 +1,4 @@
+import { PgTable } from 'drizzle-orm/pg-core'
 import { equal } from 'node:assert'
 
 import { db } from '../db/index.ts'
@@ -5,8 +6,10 @@ import * as tables from '../db/schema.ts'
 
 export async function cleanAllTables(): Promise<void> {
   await Promise.all(
-    Object.values(tables).map(table => {
-      return db.delete(table)
+    Object.values(tables).map(async table => {
+      if (table instanceof PgTable) {
+        await db.delete(table)
+      }
     })
   )
 }
